Performed Major Visual Overhaul and changed to JSX instead of TSX
There are so many changes I won't even begin to list them here, but the main two, humongous changes are the visual overhaul which is aimed to have a similar Navigation bar to the Blender Website and include a cleaner, more responsive UI, as well as switching the project from being TypeScript based to JavaScript Based. While the site itself is not fully complete, as three of the planned pages are in fact missing, the majority of the work itself is done, as the Rigs and Addons pages have gotten upgrades to ensure expandability for when I inevitably release more rigs and assets.